move area store to separate file
